+++ webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
-const PugPlugin = require('pug-plugin');
-//const PugPlugin = require('../../GitHub/webpack/pug-plugin'); // for local development only
+const HtmlBundlerPlugin = require('html-bundler-webpack-plugin');
 
 const DEFAULT_FILE = 'README.md';
 
@@ -8,26 +7,30 @@ module.exports = (env, argv) => {
   const isProd = argv.mode === 'production';
   let file = env.file || DEFAULT_FILE;
 
-  if (!file.endsWith('.md')) {
+  if (!file.endsWith('.md') && !file.endsWith('.markdown')) {
     file += '.md';
   }
 
+  const filename = path.basename(file);
   const resolvedFile = path.resolve(__dirname, file);
+  const resolvedDir = path.dirname(resolvedFile);
 
   return {
     mode: isProd ? 'production' : 'development',
-    devtool: isProd ? 'source-map' : 'inline-source-map',
+    devtool: false,
     stats: {
       preset: 'minimal',
     },
 
     resolve: {
       alias: {
-        ALIAS_TO_FILE: resolvedFile,
-        Docs: path.resolve(__dirname, 'src'),
-        Styles: path.join(__dirname, 'src/assets/styles'),
-        Scripts: path.join(__dirname, 'src/assets/scripts'),
-        Images: path.join(__dirname, 'src/assets/images'),
+        // github url to a repository
+        //'docs': 'https://github.com/webdiscus/ansis/raw/master/docs',
+
+        '@styles': path.join(__dirname, 'src/styles'),
+        '@scripts': path.join(__dirname, 'src/scripts'),
+        '@images': path.join(__dirname, 'src/images'),
+        '@npm': path.join(__dirname, 'node_modules'), // node_modules in local dir
       },
     },
 
@@ -36,90 +39,105 @@ module.exports = (env, argv) => {
       publicPath: 'auto',
     },
 
-    entry: {
-      // pass the file into Pug template
-      index: path.join(__dirname, 'src/index.pug?file=') + file,
-    },
-
     plugins: [
       // use the pug-plugin to compile pug files defined in entry
-      new PugPlugin({
+      new HtmlBundlerPlugin({
+        entry: [
+          {
+            import: 'src/index.hbs',
+            filename: 'index.html',
+            data: {
+              markdownFile: resolvedFile,
+              filePath: file,
+              title: filename,
+              markdownTheme: 'light',
+              highlightingTheme: 'prism',
+            },
+          },
+          {
+            import: 'src/index.hbs',
+            filename: 'index-dark.html',
+            data: {
+              markdownFile: resolvedFile,
+              filePath: file,
+              title: filename,
+              markdownTheme: 'dark',
+              highlightingTheme: 'prism-dark',
+            },
+          },
+        ],
         js: {
-          // output filename of scripts
-          filename: 'assets/js/[name].[contenthash:8].js',
+          filename: 'js/[name].[contenthash:8].js',
         },
         css: {
-          // output filename of styles
-          filename: 'assets/css/[name].[contenthash:8].css',
+          filename: 'css/[name].[contenthash:8].css',
         },
+
+        hotUpdate: true,
+
+        preprocessor: 'handlebars',
+
+        // resolve relative files (e.g. images) defined in *.md file
+        loaderOptions: {
+          root: path.resolve(__dirname, 'src'),
+          context: resolvedDir, // use it for html-bundler-webpack-plugin only, disable for others
+        },
+
+        watchFiles: {
+          // add external path to watch includes
+          paths: [resolvedDir],
+          includes: [
+            ///\.md/, // watch changes of *.md files in the paths
+            file, // add external *.md file to watching
+          ],
+          excludes: [
+            // exclude all files (defined as default extensions) except the exactly the file
+            new RegExp(`${resolvedDir}/(?!${filename})`)
+          ],
+        },
+
+        verbose: false,
       }),
     ],
 
     module: {
       rules: [
         {
-          test: /\.pug$/,
-          loader: PugPlugin.loader, // the pug-loader is already included in PugPlugin
-          options: {
-            method: 'render', // fast method to compile Pug files in static HTML
-            // enable filters only those used in pug
-            embedFilters: {
-              // :escape
-              escape: true,
-              // :code
-              code: {
-                className: 'language-',
-              },
-              // :highlight
-              highlight: {
-                verbose: true,
-                use: 'prismjs', // name of highlighting npm package, must be installed
-              },
-              // :markdown
-              markdown: {
-                github: true, // support github syntax for note, warning
-                highlight: {
-                  verbose: true,
-                  use: 'prismjs', // name of highlighting npm package, must be installed
-                },
-              },
-            },
-          },
-        },
-
-        // styles
-        {
-          test: /\.(css|sass|scss)$/,
+          test: /\.(s?css)$/,
           use: ['css-loader', 'sass-loader'],
         },
         {
           test: /\.(png|svg|ico)$/i,
           type: 'asset/resource',
           generator: {
-            filename: '[name].[ext]',
+            filename: '[name][ext]',
           },
         }
       ],
     },
 
     devServer: {
+      // open in default browser
+      open: true,
+
       static: {
-        // important: this is the directory of an MD file to display images with relative url
+        // important: this is the directory of an Markdown file to display images with relative url
         directory: path.dirname(resolvedFile),
       },
-      https: false,
-      compress: true,
 
       watchFiles: {
-        // watch only one file passed from commandline
-        paths: [resolvedFile],
+        // watch only one file passed from the commandline
+        paths: [resolvedFile, 'src/**/*.*'],
         options: {
           usePolling: true,
         },
       },
 
-      // open in default browser
-      open: true,
+      headers: {
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
+        "Access-Control-Allow-Headers": "X-Requested-With, content-type, Authorization"
+      },
     },
   };
 };