+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## 2.0.0 (2025-03-19)
+
+- feat: migrate to html-bundler-webpack-plugin
+- feat: add GitHub styles including Note, Warning, Important, Caution, Tip
+- feat: add switch light/dark theme
+
+## 1.1.0 (2024-05-25)
+
+- chore: upgrade the pug-plugin to v5.2.0, update dependencies
+
 ## 1.0.3 (2023-03-04)
 - chore: update Webpack config according changes in new version of pug-plugin